blankon/tests/requirements.txt
-r ../requirements.txt

pytest==8.1.1
pytest-django==3.9.0
pytest-cov==2.10.0
six==1.16.0
